% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo.R
\name{combo}
\alias{combo}
\title{combine datasets to be used in HEI scoring}
\usage{
combo(fped, diet, demograph, agethresh = 2)
}
\arguments{
\item{fped}{food pattern equivalent database data}

\item{diet}{dietary data from NHANES database}

\item{demograph}{demographic data from NHANES database}

\item{agethresh}{numeric threshold for age of survey participants to be included}
}
\value{
data frame of combined fped, dietary and demographic data
}
\description{
combine datasets to be used in HEI scoring
}
