% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_demo.R
\name{get_demo}
\alias{get_demo}
\title{retrieve demographic data from NHANES database}
\usage{
get_demo(year)
}
\arguments{
\item{year}{year combination of dataset to retrieve must be one of "2005/2006", "2007/2008", "2009/2010" or "2011/2012"}
}
\value{
data frame of NHANES demographic database
}
\description{
retrieve demographic data from NHANES database
}

