% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diet.R
\name{get_diet}
\alias{get_diet}
\title{retrieve dietary data from NHANES database}
\usage{
get_diet(x, strict = TRUE)
}
\arguments{
\item{x}{dataset to retrieve}

\item{strict}{indicator as to whether or not all variables should be returned}
}
\value{
data frame of NHANES dietary database
}
\description{
retrieve dietary data from NHANES database
}

