% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fped.R
\name{get_fped}
\alias{get_fped}
\title{retrieve FPED data}
\usage{
get_fped(year, day = "both")
}
\arguments{
\item{year}{year combination of dataset to retrieve must be one of "2005/2006", "2007/2008", "2009/2010", "2011/2012" or "2013/2014"}

\item{day}{identifier for FPED data to retrieve ... must be one of "first", "second" or "both" (default is set to "both")}
}
\value{
object of class \code{data.frame} containing FPED data for specificed year and day
}
\description{
retrieve FPED data
}
