% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leg_all.R
\name{leg_all}
\alias{leg_all}
\title{allocate legumes for HEI scoring}
\usage{
leg_all(dat)
}
\arguments{
\item{dat}{data to be processed}
}
\value{
data frame of dietary values with legumes allocated appropriately
}
\description{
allocate legumes for HEI scoring
}
