% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_demo.R
\name{get_demo}
\alias{get_demo}
\title{retrieve demographic data from NHANES database}
\usage{
get_demo(x, strict = TRUE)
}
\arguments{
\item{x}{dataset to retrieve}

\item{strict}{indicator as to whether or not all variables should be returned}
}
\value{
data frame of NHANES demographic database
}
\description{
retrieve demographic data from NHANES database
}

