% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diet.R
\name{get_diet}
\alias{get_diet}
\title{retrieve dietary data from NHANES database}
\usage{
get_diet(year, day = "both")
}
\arguments{
\item{year}{year combination of dataset to retrieve must be one of "2005/2006", "2007/2008", "2009/2010" or "2011/2012"}

\item{day}{identifier for survey day to retrive ... must be one of "first", "second" or "both" (default is set to "both")}
}
\value{
data frame of NHANES dietary database
}
\description{
retrieve dietary data from NHANES database
}
