% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hei.R
\name{hei}
\alias{hei}
\title{Calculate Healthy Eating Index (HEI) 2015 scores:
total and components 1 through 13}
\usage{
hei(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
data.frame
}
\description{
Calculate Healthy Eating Index (HEI) 2015 scores:
total and components 1 through 13
}
