% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hei.R
\name{hei}
\alias{hei}
\title{Individual HEI scores calculation.}
\usage{
hei(dat)
}
\arguments{
\item{dat}{data to be scored}
}
\value{
Object of class \code{data.frame} containing all columns of input data set as well as 33 columns of calculated data related to HEI scoring and, significantly, a 70th column containing the total HEI score for each participant.
}
\description{
\code{hei} takes a data set with columns of formatted dietary data (per \code{hei::combo}) and returns a data set with calculated total HEI scores for each individual along with scores for HEI constituent scoring categories.
}
\examples{
\dontrun{
fped0910 <- get_fped("2009/2010", "both")
diet0910 <- get_diet("2009/2010", "both")
demo0910 <- get_demo("2009/2010")
all0910 <- combo(fped0910,diet0910,demo0910)
hei(all0910)

fped0506 <- get_fped("2005/2006", "first")
diet0506 <- get_diet("2005/2006", "first")
demo0506 <- get_demo("2005/2006")
all0506 <- combo(fped0506,diet0506,demo0506)
hei(all0506)
}
}
\references{
\url{https://www.cnpp.usda.gov/healthyeatingindex}
}
