% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hei.R
\name{hei}
\alias{hei}
\title{Individual HEI scores calculation.}
\usage{
hei(dat)
}
\arguments{
\item{dat}{data to be scored}
}
\value{
Object of class \code{data.frame} containing all columns of input data set as well as 33 columns of calculated data related to HEI scoring and, significantly, a 70th column containing the total HEI score for each participant.
}
\description{
\code{hei} takes a data set with columns of formatted dietary data (per \code{hei::combo}) and returns a data set with calculated total HEI scores for each individual along with scores for HEI constituent scoring categories.
}
\examples{
hei(combo(get_fped("2009/2010"), get_diet("2009/2010"), get_demo("2009/2010")))
hei(combo(get_fped("2009/2010", "first"), get_diet("2009/2010", "first"), get_demo("2009/2010")))
hei(combo(get_fped("2009/2010", "first" _diet("2009/2010", "first"), get_demo("2009/2010"), 80))
}
